backend/requirements.txt
asgiref==3.2.3
astroid==2.3.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-location-field==2.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.4.2
Markdown==3.1.1
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
Pillow==7.0.0
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
